requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
cytoolz==0.12.0
defi-head-core @ git+https://github.com/e183b796621afbf902067460/defi-head-core.git@71c63885d3dec58b92087ace732e65b16e6bb45b
ecdsa==0.18.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.0.4
fastapi==0.88.0
fastapi-utils==0.2.1
Flask==2.2.2
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
hexbytes==0.3.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
passlib==1.7.4
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
protobuf==3.19.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.16.0
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
pyseto==1.6.12
pytest==7.2.0
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rlp==2.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
starlette==0.22.0
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
varint==1.0.2
web3==5.31.3
websockets==9.1
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.11.0
